Implemented 'Bad Rating' logic
API call and deserialising into objects, mapping these into various objects, and using on the UI for user search. Implemented some basic rating functionality in RatingCalculation.cs which just finds the mean average for a rating. This is WIP to  eventually use weighted averages based on how much a user plays the champion as well as it's standard champion rating.